String "keys" lead to inconsistent results:
Consider ensuring that the methods accepting a string "key" accept only
a hexadecimal string and then decode that string directly to the bytes
said hexadecmial characters represent.  Or possibly consider removing
the Encode/Decode methods that accept a string parameter for a key.

HMACSHA256 algorithms generally take an array of bytes as the data to
be hashed as well as another array of bytes as the key.

In this JWT implementation the Encoding.UTF8.GetBytes(key) call will
simply convert the string to its UTF8 character bytes.  For example,

Encoding.UTF8.GetBytes("0123456789abcdef") ==
	byte[]
	{
		0x30, 0x31, 0x32, 0x33,
		0x34, 0x35, 0x36, 0x37,
		0x38, 0x39, 0x61, 0x62,
		0x63, 0x64, 0x65, 0x66 // 16 bytes
	}

and

Encoding.UTF8.GetBytes("0123456789ABCDEF") ==
	byte[]
	{
		0x30, 0x31, 0x32, 0x33,
		0x34, 0x35, 0x36, 0x37,
		0x38, 0x39, 0x41, 0x42,
		0x43, 0x44, 0x45, 0x46 // 16 bytes
	}

Where as

"0123456789abcdef".FromHex() ==
	byte[]
	{
		0x01, 0x23, 0x45, 0x67, 0x89, 0xab, 0xcd, 0xef // only 8 bytes
	}

and

"0123456789ABCDEF".FromHex() ==
	byte[]
	{
		0x01, 0x23, 0x45, 0x67, 0x89, 0xab, 0xcd, 0xef // only 8 bytes
	}

Not only does the Encoding.UTF8.GetBytes() method above lead to double
the intended length byte array it also makes the hexadecimal string
value case sensitive:   A hexadecimal "a" and a hexadecimal "A"
both represent a decimal of "10".  However, a UTF8 character "a"
represents a hexadecimal value of 0x61 whereas a UTF8 character "A"
represents a hexadecimal value of 0x41.